refactor: String 내장메서드 padStart 적용
