+++ src/components/shared/NDigitBall/style.css
@@ -1,4 +1,4 @@
-.TwoDigitBall {
+.NDigitBall {
   display: inline-block;
   width: 1rem;
   height: 1rem;
