+++ src/components/shared/index.js
@@ -3,7 +3,7 @@ import PlainButton from './PlainButton';
 import SubmitButton from './SubmitButton';
 import ToggleButton from './ToggleButton';
 import XButton from './XButton';
-import TwoDigitBall from './TwoDigitBall';
+import TwoDigitBall from './NDigitBall';
 import LottoBall from './LottoBall';
 
 export { Animation, PlainButton, SubmitButton, ToggleButton, XButton, TwoDigitBall, LottoBall };
